requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
google-generativeai==0.3.2
huggingface-hub>=0.26.0
requests==2.31.0
aiohttp==3.9.1
